Refactor root layout and extract filter toolbar
